+++ gnssrefl/gps.py
@@ -207,6 +207,7 @@ def define_and_xz_snr(station,year,doy,snr):
         else:
             if os.path.isfile(fname3):
                 subprocess.call(['gunzip', fname3])
+                fname2 = fname3 #Switch file names for .xz to .gz, for returning proper name below
         # make sure the uncompression worked
         if os.path.isfile(fname):
             snre = True
@@ -823,7 +824,7 @@ def orbfile_cddis(name, year, secure_file, secure_dir, file2):
     """
     # assume file was not found
     foundit = False
-    print(secure_dir, secure_file)
+    #print(secure_dir, secure_file)
     # Z or gz expected ...
     if secure_file[-3:] == '.gz':
         gzip = True
@@ -3848,8 +3849,9 @@ def snr_exist(station,year,doy,snrEnd):
         snre = True # but needs to be uncompressed
         subprocess.call(['unxz', fname2])
     if os.path.isfile(fname3) and (not snre):
-            snre = True # but needs to be ungzipped 
-            subprocess.call(['gunzip', fname3])
+        snre = True # but needs to be ungzipped 
+        #subprocess.call(['gunzip', fname3])
+        #TS - Removed this line Aug 2023 to stop unecessary decompression in nmea2snr
 
     return snre 
 
@@ -6004,10 +6006,10 @@ def save_plot(plotname):
     Parameters
     ----------
     plotname : str
-        name of png file
+        name of output figure file
     """
     plt.savefig(plotname,dpi=300)
-    print('png file saved as: ', plotname)
+    print('Plot file saved as: ', plotname)
 
 
 def make_azim_choices(alist):
@@ -6356,4 +6358,40 @@ def gbm_orbits_direct(year,month,day):
     #    print('Orbit found')
 
     return return_name, fdir, foundit
+
+
+def checkFiles(station, extension):
+    """
+    apparently no one consistently checks for the Files directory existence.
+    this is an attempt to fix that.
+
+    Parameters
+    ----------
+    station : str
+        4 ch station ID
+
+    extension : str
+        subdirectory for results in $REFL_CODE/Files/station
+
+    """
+    xdir = os.environ['REFL_CODE']
+    if not os.path.isdir(xdir):
+        print('REFL_CODE environment variable has not been set. Exiting')
+        sys.exit()
+
+    txtdir = xdir + '/Files/' 
+    if not os.path.exists(txtdir) :
+        subprocess.call(['mkdir', txtdir])
+        print(txtdir , ' has been created.')
+    txtdir = xdir + '/Files/'  + station
+    if not os.path.exists(txtdir) :
+        subprocess.call(['mkdir', txtdir])
+        print(txtdir , ' has been created.')
+
+    if (len(extension) > 0):
+        txtdir = xdir + '/Files/'  + station + '/' + extension
+        if not os.path.exists(txtdir) :
+            subprocess.call(['mkdir', txtdir])
+            print(txtdir , ' has been created')
+