+++ gnssrefl/quickLook_function.py
@@ -14,7 +14,7 @@ import gnssrefl.rinex2snr as rinex
 
 
 def quickLook_function(station, year, doy, snr_type,f,e1,e2,minH,maxH,reqAmp,pele,satsel,PkNoise,fortran,
-        pltscreen,azim1,azim2,ediff, delTmax,**kwargs):
+        pltscreen,azim1,azim2,ediff, delTmax, hires_figs,**kwargs):
     """
     This is the main function to compute spectral characteristics of a SNR file.
     It takes in all user inputs and calculates reflector heights. It makes two png files to summarize
@@ -60,6 +60,8 @@ def quickLook_function(station, year, doy, snr_type,f,e1,e2,minH,maxH,reqAmp,pel
          QC parameter - restricts length of arcs (degrees)
     delTmax : float
          maximum arc length in minutes
+    hires_figs : bool
+         whether to use eps instead of png
 
     """
 
@@ -299,11 +301,14 @@ def quickLook_function(station, year, doy, snr_type,f,e1,e2,minH,maxH,reqAmp,pel
         plt.suptitle(tt, fontsize=FS)
             # if you have no results, no point plotting them!
         if (allpoints > 0):
-            filename = fdir + '/quickLook_lsp.png'
-            print('plot saved to ', filename)
+            if hires_figs:
+                filename = fdir + '/quickLook_lsp.eps'
+            else:
+                filename = fdir + '/quickLook_lsp.png'
+            print('Plot saved to ', filename)
             plt.savefig(filename)
         # make second plot
-            goodbad(quicklog,station,year,doy,minH,maxH,PkNoise,reqAmp,f,e1,e2)
+            goodbad(quicklog,station,year,doy,minH,maxH,PkNoise,reqAmp,f,e1,e2,hires_figs)
         else:
             print('You made a selection that does not exist (i.e. frequency or satellite or constellation)')
 
@@ -318,7 +323,7 @@ def quickLook_function(station, year, doy, snr_type,f,e1,e2,minH,maxH,reqAmp,pel
 
     return data,datakey
 
-def goodbad(fname,station,year,doy,h1,h2,PkNoise,reqAmp,freq,e1,e2):
+def goodbad(fname,station,year,doy,h1,h2,PkNoise,reqAmp,freq,e1,e2,hires_figs):
     """
     makes a plot that shows "good" and "bad" refletor height retrievals as a 
     function of azimuth
@@ -347,6 +352,8 @@ def goodbad(fname,station,year,doy,h1,h2,PkNoise,reqAmp,freq,e1,e2):
         minimum elevation angle (deg)
     e2 : float
         maximum elevation angle (deg)
+    hires_figs: bool
+        whether you want eps instead of png
 
     Returns
     -------
@@ -415,8 +422,12 @@ def goodbad(fname,station,year,doy,h1,h2,PkNoise,reqAmp,freq,e1,e2):
 
     # existence of the output directory is checked earlier
     fdir = os.environ['REFL_CODE'] + '/Files/' + station 
-    f = fdir + '/quickLook_summary.png'
-    print('plot saved to ', f)
+    if hires_figs :
+        f = fdir + '/quickLook_summary.eps'
+    else:
+        f = fdir + '/quickLook_summary.png'
+
+    print('Plot saved to ', f)
     plt.savefig(f)
 
 def read_snr_simple(obsfile):