+++ gnssrefl/subdaily.py
@@ -78,6 +78,8 @@ def mirror_plot(tnew,ynew,spl_x,spl_y,txtdir,station,beginT,endT):
         first time (day of year) real RH measurement
     endT : float
         last time (day of year) for first real RH measurement
+    extension : str
+        name of testing strategy (used in gnssir)
 
     """
     fig=plt.figure(figsize=(10,4))
@@ -848,9 +850,13 @@ def stack_two_more(otimes,tv,ii,jj,stats, station, txtdir, sigma,kplt,hires_figs
     plt.gca().invert_yaxis()
     plt.legend(loc="best")
     plt.grid()
-    plotname = txtdir + '/' + station + '_outliers_wrt_az.png'
+    if hires_figs:
+        plotname = txtdir + '/' + station + '_outliers_wrt_az.eps'
+    else:
+        plotname = txtdir + '/' + station + '_outliers_wrt_az.png'
+
     plt.savefig(plotname,dpi=300)
-    print('png file saved as: ', plotname)
+    print('Plot file saved as: ', plotname)
 
     #    fig=plt.figure(figsize=(10,6))
     fig = plt.figure(figsize=(10,6))
@@ -885,15 +891,15 @@ def stack_two_more(otimes,tv,ii,jj,stats, station, txtdir, sigma,kplt,hires_figs
     plt.xticks(rotation =45,fontsize=8); plt.yticks(fontsize=8)
     plt.title('Edited ' + station.upper() + ' Reflector Heights', fontsize=8)
     plt.grid() ; fig.autofmt_xdate()
-    plotname = txtdir + '/' + station + '_outliers.png'
     if hires_figs:
-        plotname = txtdir + '/' + station + '_outliers.png'
+        plotname = txtdir + '/' + station + '_outliers.eps'
         plt.savefig(plotname,dpi=300)
     else:
+        plotname = txtdir + '/' + station + '_outliers.png'
         plt.savefig(plotname,dpi=300)
 
     plt.ylim((savey1, savey2))
-    print('png file saved as: ', plotname)
+    print('Plot file saved as: ', plotname)
     if kplt:
         fig = plt.figure()
         ax1 = fig.add_subplot(211)
@@ -999,7 +1005,7 @@ def apply_new_constraints(tv,azim1,azim2,ampl,peak2noise,d1,d2,h1,h2):
     return tv,t,rh,firstdoy,lastdoy
 
 
-def rhdot_plots(th,correction,rhdot_at_th, tvel,yvel,fs,station,txtdir):
+def rhdot_plots(th,correction,rhdot_at_th, tvel,yvel,fs,station,txtdir,hires_figs):
     """
     make rhdot correction plots
 
@@ -1021,6 +1027,8 @@ def rhdot_plots(th,correction,rhdot_at_th, tvel,yvel,fs,station,txtdir):
         station name
     txtdir : str
         file directory for output
+    hires_figs : bool
+        whether you want eps instead of png 
 
     """
     fig=plt.figure(figsize=(10,6))
@@ -1042,7 +1050,10 @@ def rhdot_plots(th,correction,rhdot_at_th, tvel,yvel,fs,station,txtdir):
     plt.title('surface velocity')
     plt.ylabel('meters/hour'); plt.xlabel('days of the year')
     plt.xlim((A1,A2))
-    g.save_plot(txtdir + '/' + station + '_rhdot3.png')
+    if hires_figs:
+        g.save_plot(txtdir + '/' + station + '_rhdot3.eps')
+    else:
+        g.save_plot(txtdir + '/' + station + '_rhdot3.png')
 
 
 
@@ -1235,7 +1246,7 @@ def rhdot_correction2(station,fname,fname_new,pltit,outlierV,outlierV2,**kwargs)
         print('using knots_test')
 
     print('\n>>>>>>>>>>>>>>>>>>>> Entering second section of subdaily code <<<<<<<<<<<<<<<<<<<<<<<<')
-    print('\nComputes rhdot correction and bias correction for subdaily')
+    print('\nComputes rhdot correction and interfrequency bias correction for subdaily')
     print('\nInput filename:', fname)
     print('\nOutput filename: ', fname_new)
 
@@ -1310,8 +1321,16 @@ def rhdot_correction2(station,fname,fname_new,pltit,outlierV,outlierV2,**kwargs)
     yvel = obsPerHour*np.diff(spl_y)
 
     rhdot_at_th = np.interp(th, tvel, yvel)
+    # RH dot correction
     correction = xfac*rhdot_at_th
-    correctedRH = h-correction
+
+    # 2023 august 25
+    if not apply_rhdot :
+        print('You requested RHdot correction NOT be applied.')
+        correctedRH = h
+    else:
+    # corrected reflector height
+        correctedRH = h-correction
 
     # this is RH with the RHdot correction
     residual_before = h - spl_at_GPS_times
@@ -1378,14 +1397,17 @@ def rhdot_correction2(station,fname,fname_new,pltit,outlierV,outlierV2,**kwargs)
     plt.legend(loc="upper left")
     plt.xlim((plot_begin, plot_end))
 
-    g.save_plot(txtdir + '/' + station + '_rhdot2.png')
+    if hires_figs:
+        g.save_plot(txtdir + '/' + station + '_rhdot2.eps')
+    else:
+        g.save_plot(txtdir + '/' + station + '_rhdot2.png')
 
 
     # update the residual vector as well
     residual_after = residual_after[ii]
 
     # make the RHdot plot as well
-    rhdot_plots(th,correction,rhdot_at_th, tvel,yvel,fs,station,txtdir)
+    rhdot_plots(th,correction,rhdot_at_th, tvel,yvel,fs,station,txtdir,hires_figs)
 
     writecsv = False ; extraline = ''
     # write out the new solutions with RHdot and without 3 sigma outliers
@@ -1510,6 +1532,7 @@ def rhdot_correction2(station,fname,fname_new,pltit,outlierV,outlierV2,**kwargs)
     plt.title('Station: ' + station + ', new spline, RHdot corr/InterFreq corr/outliers removed')
     plt.xlabel('days of the year')
     # put hires_figs boolean here
+    print(txtdir)
     if hires_figs:
         g.save_plot(txtdir + '/' + station + '_rhdot4.eps')
     else:
@@ -1529,7 +1552,11 @@ def rhdot_correction2(station,fname,fname_new,pltit,outlierV,outlierV2,**kwargs)
     plt.legend(loc="upper left")
 
     print('RMS with frequency biases and RHdot taken out (m) ', np.round(newsigma,3)  )
-    g.save_plot(txtdir + '/' + station + '_rhdot5.png')
+    if hires_figs:
+        g.save_plot(txtdir + '/' + station + '_rhdot5.eps')
+    else:
+        g.save_plot(txtdir + '/' + station + '_rhdot5.png')
+
     plt.close() # dont send this one to the screen
 
     # bias corrected - and again without 3 sigma outliers
@@ -1780,5 +1807,5 @@ def numsats_plot(station,tval,nval,Gval,Rval,Eval,Cval,txtdir,fs,hires_figs):
         plt.savefig(plotname,dpi=300)
     else:
         plt.savefig(plotname,dpi=300)
-    print('png file saved as: ', plotname)
+    print('Plot file saved as: ', plotname)